send training-cancelled mail to registered users

+++ src/app/(main)/trainer/actions.ts
@@ -3,7 +3,9 @@
 import { revalidatePath } from "next/cache";
 import { z } from "zod";
 
+import { formatTrainingDate } from "@/lib/date";
 import { getServerSession } from "@/lib/next-auth";
+import { sendTrainingCancelledMail } from "@/lib/postmark";
 import { prisma } from "@/lib/prisma";
 
 export async function getMyTrainings() {
@@ -31,11 +33,42 @@ export async function getMyTrainings() {
 }
 
 export async function deleteTraining(id: string) {
+  const training = await prisma.training.findFirst({
+    where: {
+      id,
+    },
+    include: {
+      author: true,
+      registrations: {
+        include: {
+          user: true,
+        },
+      },
+    },
+  });
+
+  if (!training) {
+    return {
+      error: "training not found",
+    };
+  }
+
   await prisma.training.delete({
     where: {
       id,
     },
   });
+
+  const registeredUsers = training.registrations.map((r) => r.user.email);
+
+  registeredUsers.forEach((email) => {
+    sendTrainingCancelledMail({
+      to: email,
+      trainer: training.author.name || "Ein:e Trainer:in",
+      date: formatTrainingDate(training.start, training.end),
+    });
+  });
+
   revalidatePath("/trainer");
 }
